Change mock data structure for movies. Refactor. Fix problems
