+++ src/components/profile-rank.js
@@ -1,4 +1,4 @@
-import {getProfileRank} from "../mock/rank/rank";
+import {getProfileRank} from "../mock/rank";
 
 /**
  * Создание разметки блока ранга профиля пользователя
