+++ src/mock/comments/comment-consts.js
@@ -1,0 +1,19 @@
+export const COMMENT_AUTHORS = [
+  `Neikan`,
+  `Mictlantechuhtli`,
+  `Эрнанда`,
+  `Sorpi`,
+  `БегущаяПоГраблям`,
+  `Ксардасс`,
+  `Azatot`
+];
+
+export const COMMENT_TEXTS = [
+  `Начну с главного — это великолепно! Настолько, что я даже решился написать отзыв, как умею.`,
+  `Возникает вопрос: а на кого рассчитан этот фильм? Его нельзя показывать из-за наличия жестокости. Откровенно глупо и безсмысленно: эротики нет, романтических нет, глубокий смысл, экшен и драматизм отсутствуют!`,
+  `Когда я увидел анонс картины, я очень заинтересовался и решил обязательно посмотреть! Это круто! 100 из 10!`,
+  `По большому счету фильм — освежающий взгляд на известного героя. К сожалению, взгляд хотя и свежий, но всё же пустой. Как и многие другие этот фильм — необходимое отклонение от канона, но зрелище вышло больно странным и эксцентричным из-за самого подхода к созданию.`,
+  `Честно сказать сначала мне было интересно. Но потом... Жаль потраченного времени`,
+  `Это безусловно один из самых необычных и неожиданных продуктов вселенной DC. Очень смелый, неожиданный и оригинальный эксперимент, который придал глоток свежего воздуха во вселенную Бэтмена. Во многом, весьма необычным сеттингом и потрясающей режессурой, ради которой и стоит посмотреть данный фильм и нисколько не пожалеть о потраченном времени.`,
+  `Увидев трейлер в сети этой картины, сразу занес в список ожиданий, так был удивлен! И вот недавно вышел в сеть это фильм, но впечатления от просмотра получил двоякие слишком(`
+];
