+++ src/mock/films/film-consts.js
@@ -1,0 +1,198 @@
+export const COUNT_COMMENTS_MAX = 5;
+export const RAITING_MAX = 10;
+
+export const TITLES = [
+  {
+    translate: `Бэтмен`,
+    original: `Batman`
+  },
+  {
+    translate: `Бэтмен возвращается`,
+    original: `Batman Returns`
+  },
+  {
+    translate: `Бэтмен навсегда`,
+    original: `Batman Forever`
+  },
+  {
+    translate: `Бэтмен и Робин`,
+    original: `Batman & Robin`
+  },
+  {
+    translate: `Бэтмен: Начало`,
+    original: `Batman Begins`
+  },
+  {
+    translate: `Темный рыцарь`,
+    original: `The Dark Knight`
+  },
+  {
+    translate: `Темный рыцарь: Возрождение легенды`,
+    original: `The Dark Knight Rises`
+  },
+  {
+    translate: `Бэтмен-ниндзя`,
+    original: `Batman Ninja`
+  },
+  {
+    translate: `Бэтмен против Супермена: На заре справедливости`,
+    original: `Batman v Superman: Dawn of Justice`
+  },
+  {
+    translate: `Отряд самоубийц`,
+    original: `Suicide Squad`
+  },
+  {
+    translate: `Лига справедливости`,
+    original: ``
+  },
+  {
+    translate: `Джокер`,
+    original: `Joker`
+  },
+  {
+    translate: `Безграничный Бэтмен`,
+    original: `Batman Unlimited`
+  },
+  {
+    translate: `Лего Фильм: Бэтмен`,
+    original: `The Lego Batman Movie`
+  },
+  {
+    translate: `Готэм`,
+    original: `Gotham`
+  }
+];
+
+export const POSTERS = [
+  `made-for-each-other.png`,
+  `popeye-meets-sinbad.png`,
+  `sagebrush-trail.jpg`,
+  `santa-claus-conquers-the-martians.jpg`,
+  `the-dance-of-life.jpg`,
+  `the-great-flamarion.jpg`,
+  `the-man-with-the-golden-arm.jpg`,
+];
+
+export const DIRECTORS = [
+  `Тим Бертон`,
+  `Джоэл Шумахер`,
+  `Кристофер Нолан`,
+  `Джей Олива`,
+  `Крис МакКей`,
+  `Зак Снайдер`,
+  `Дзюнпэй Мидзусаки`
+];
+
+export const SCREEN_WRITERS = [
+  `Акива Голдсман`,
+  `Боб Кейн`,
+  `Кристофер Нолан`,
+  `Джонатан Нолан`,
+  `Дэвид С. Гойер`,
+  `Боб Кейн`,
+  `Сет Грэм-Смит`,
+  `Крис МакКенна`,
+  `Скотт Снайдер`,
+  `Сэм Хэмм`,
+];
+
+export const ACTORS = [
+  `Майкл Китон`,
+  `Джек Николсон`,
+  `Ким Бейсингер`,
+  `Роберт Вул`,
+  `Кристиан Бэйл`,
+  `Том Харди`,
+  `Энн Хэтэуэй`,
+  `Джозеф Гордон-Левитт`,
+  `Марион Котийяр`,
+  `Арнольд Шварценеггер`,
+  `Джордж Клуни`,
+  `Крис О’Доннелл`,
+  `Ума Турман`,
+  `Коити Ямадэра`,
+  `Ватару Такаги`,
+  `Риэ Кугимия`,
+  `Аи Какума`,
+  `Хотю Оцука`,
+  `Такэхито Коясу`,
+  `Кэнта Миякэ`
+];
+
+export const COUNTRIES = [
+  `France`,
+  `Germany`,
+  `India`,
+  `Italy`,
+  `Japan`,
+  `Russia`,
+  `Spain`,
+  `UK`,
+  `USA`
+];
+
+export const GENRES = [
+  `Aниме`,
+  `Биография`,
+  `Боевик`,
+  `Вестерн`,
+  `Военный`,
+  `Детектив`,
+  `Детский`,
+  `Документальный`,
+  `Драма`,
+  `Исторический`,
+  `Кинокомикс`,
+  `Комедия`,
+  `Концерт`,
+  `Короткометражный`,
+  `Криминал`,
+  `Мелодрама`,
+  `Мистика`,
+  `Мультфильм`,
+  `Мюзикл`,
+  `Научный`,
+  `Нуар`,
+  `Приключения`,
+  `Семейный`,
+  `Сериал`,
+  `Спорт`,
+  `Триллер`,
+  `Ужасы`,
+  `Фантастика`,
+  `Фэнтези`,
+  `Эротика`
+];
+
+export const AGE_RATINGS = [
+  `0+`,
+  `6+`,
+  `12+`,
+  `16+`,
+  `18+`,
+];
+
+export const DESCRIPTION = [
+  `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
+  `Cras aliquet varius magna, non porta ligula feugiat eget.`,
+  `Fusce tristique felis at fermentum pharetra.`,
+  `Aliquam id orci ut lectus varius viverra.`,
+  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
+  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
+  `Sed sed nisi sed augue convallis suscipit in sed felis.`,
+  `Aliquam erat volutpat.`,
+  `Nunc fermentum tortor ac porta dapibus.`,
+  `In rutrum ac purus sit amet tempus.`
+];
+
+
+export const CountDescription = {
+  MIM: 1,
+  MAX: 5
+};
+
+export const CountDuration = {
+  HOURS_MAX: 3,
+  MINUTES_MIN: 59,
+};
